requirements.txt
yfinance
pandas
numpy
streamlit
matplotlib
scikit-learn
textblob
openai
python-dotenv
ta